requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
PyMySQL==1.1.0
SQLAlchemy==2.0.28
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
gunicorn
PyMySQL[rsa]
python-dotenv








